Registration flow update
