+++ grails-app/controllers/com/secured/main/UrlMappings.groovy
@@ -10,8 +10,8 @@ class UrlMappings {
             }
         }
         "/api"(controller: "apiTest")
-        "/register"(controller: 'user',action: 'register')
-        "/register/confirm"(controller: 'user', action: "confirm")
+        "/register"(controller: 'register',action: 'register')
+        "/register/confirm"(controller: 'register', action: "confirm")
         "/"(controller: 'main', action:'home')
         "/verify"(controller: "main", action: "verify")
         "500"(view:'/error')
