+++ grails-app/services/com/secured/auth/UserValidatorService.groovy
@@ -1,6 +1,6 @@
 package com.secured.auth
 
-import com.secured.auth.commands.UserCommand
+import com.secured.user.UserCommand
 import grails.gorm.transactions.Transactional
