+++ src/test/groovy/com/secured/auth/UserControllerSpec.groovy
@@ -1,5 +1,6 @@
 package com.secured.auth
 
+import com.secured.main.UserController
 import grails.testing.web.controllers.ControllerUnitTest
 import spock.lang.Specification
